requirements.txt
tortoise-orm[asyncpg]
fastapi
pydantic
uvicorn
aerich
pydantic-partial

git+https://github.com/HexChap/MSCore#egg=ms_core